# backend/openapi.yaml
openapi: 3.0.3
info:
  title: FinTrak
  description: |
    Personal finance tracker API. All API endpoints are prefixed with `/api/v1`.

    ## Authentication
    Protected routes require a Bearer token in the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```

    Tokens are obtained via `/auth/register` or `/auth/login`.
  version: 0.1.0-alpha.0
  contact:
    name: FinTrak
  license:
    name: AGPL v3.0
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
paths:
...
  /rules:
    get:
      tags:
        - Rules
      summary: List all rules
      operationId: getRules
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rule"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Rules
      summary: Create a new rule
      operationId: createRule
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRuleRequest"
      responses:
        "201":
          description: Rule created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /rules/{id}:
    put:
      tags:
        - Rules
      summary: Update a rule
      operationId: updateRule
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRuleRequest"
      responses:
        "200":
          description: Rule updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags:
        - Rules
      summary: Delete a rule
      operationId: deleteRule
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Rule deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /rules/apply:
    post:
      tags:
        - Rules
      summary: Apply all rules to uncategorized transactions
      operationId: applyRules
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Rules applied successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ============= Links =============
...
components:
  schemas:
...
    CreateRuleRequest:
      type: object
      required:
        - pattern
        - categoryId
      properties:
        pattern:
          type: string
        matchType:
          type: string
          enum: [contains, starts_with, exact]
          default: contains
        categoryId:
          type: string
          format: uuid
        payeeId:
          type: string
          format: uuid
        priority:
          type: integer
    UpdateRuleRequest:
      type: object
      properties:
        pattern:
          type: string
        matchType:
          type: string
          enum: [contains, starts_with, exact]
        categoryId:
          type: string
          format: uuid
        payeeId:
          type: string
          format: uuid
        priority:
          type: integer
...
    Rule:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pattern:
          type: string
        matchType:
          type: string
          enum: [contains, starts_with, exact]
        categoryId:
          type: string
          format: uuid
        categoryName:
          type: string
        payeeId:
          type: string
          format: uuid
        payee:
          type: string
        priority:
          type: integer
...
    ErrorResponse:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          minItems: 1
          items:
            type: object
            required:
              - message
            properties:
              field:
                type: string
              tag:
                type: string
              message:
                type: string
    MessageResponse:
      type: object
      properties:
        message:
          type: string
  responses:
    BadRequest:
      description: Bad request (invalid input)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/register or /auth/login